Refactor SetOperator and child classes for cleaner abstraction (#16602)
