+++ pinot-query-runtime/src/main/java/org/apache/pinot/query/runtime/operator/set/MinusAllOperator.java
@@ -16,11 +16,12 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.query.runtime.operator;
+package org.apache.pinot.query.runtime.operator.set;
 
 import java.util.List;
 import org.apache.pinot.common.utils.DataSchema;
 import org.apache.pinot.core.data.table.Record;
+import org.apache.pinot.query.runtime.operator.MultiStageOperator;
 import org.apache.pinot.query.runtime.plan.OpChainExecutionContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -29,7 +30,7 @@ import org.slf4j.LoggerFactory;
 /**
  * EXCEPT ALL operator.
  */
-public class MinusAllOperator extends SetOperator {
+public class MinusAllOperator extends RightRowSetBasedSetOperator {
   private static final Logger LOGGER = LoggerFactory.getLogger(MinusAllOperator.class);
   private static final String EXPLAIN_NAME = "MINUS_ALL";
