+++ pinot-query-runtime/src/main/java/org/apache/pinot/query/runtime/operator/set/SetOperator.java
@@ -16,20 +16,16 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.query.runtime.operator;
+package org.apache.pinot.query.runtime.operator.set;
 
 import com.google.common.base.Preconditions;
-import com.google.common.collect.HashMultiset;
-import com.google.common.collect.Multiset;
-import java.util.ArrayList;
 import java.util.List;
 import org.apache.pinot.common.datatable.StatMap;
 import org.apache.pinot.common.utils.DataSchema;
 import org.apache.pinot.core.common.ExplainPlanRows;
-import org.apache.pinot.core.data.table.Record;
 import org.apache.pinot.core.operator.ExecutionStatistics;
 import org.apache.pinot.query.runtime.blocks.MseBlock;
-import org.apache.pinot.query.runtime.blocks.RowHeapDataBlock;
+import org.apache.pinot.query.runtime.operator.MultiStageOperator;
 import org.apache.pinot.query.runtime.plan.OpChainExecutionContext;
 import org.apache.pinot.segment.spi.IndexSegment;
 
@@ -42,7 +38,6 @@ import org.apache.pinot.segment.spi.IndexSegment;
  * UnionOperator: The right child operator is consumed in a blocking manner.
  */
 public abstract class SetOperator extends MultiStageOperator {
-  protected final Multiset<Record> _rightRowSet;
 
   protected final MultiStageOperator _leftChildOperator;
   protected final MultiStageOperator _rightChildOperator;
@@ -59,7 +54,6 @@ public abstract class SetOperator extends MultiStageOperator {
     Preconditions.checkState(inputOperators.size() == 2, "Set operator should have 2 child operators");
     _leftChildOperator = inputOperators.get(0);
     _rightChildOperator = inputOperators.get(1);
-    _rightRowSet = HashMultiset.create();
     _isRightChildOperatorProcessed = false;
   }
 
@@ -123,67 +117,14 @@ public abstract class SetOperator extends MultiStageOperator {
     }
   }
 
-  /**
-   * Processes the right child operator and build the set of rows that can be used to filter the left child. This method
-   * can be overridden to also be able to return data blocks while processing the right operator.
-   *
-   * @return either a data block containing rows or an EoS block, never {@code null}.
-   */
-  protected MseBlock processRightOperator() {
-    MseBlock block = _rightChildOperator.nextBlock();
-    while (block.isData()) {
-      MseBlock.Data dataBlock = (MseBlock.Data) block;
-      for (Object[] row : dataBlock.asRowHeap().getRows()) {
-        _rightRowSet.add(new Record(row));
-      }
-      sampleAndCheckInterruption();
-      block = _rightChildOperator.nextBlock();
-    }
-    assert block.isEos();
-    return block;
-  }
+  protected abstract MseBlock processLeftOperator();
 
-  /**
-   * Processes the left child operator and returns blocks of rows that match the criteria defined by the set operation.
-   *
-   * @return block containing matched rows or EoS, never {@code null}.
-   */
-  protected MseBlock processLeftOperator() {
-    // Keep reading the input blocks until we find a match row or all blocks are processed.
-    // TODO: Consider batching the rows to improve performance.
-    while (true) {
-      MseBlock leftBlock = _leftChildOperator.nextBlock();
-      if (leftBlock.isEos()) {
-        return leftBlock;
-      }
-      MseBlock.Data dataBlock = (MseBlock.Data) leftBlock;
-      List<Object[]> rows = new ArrayList<>();
-      for (Object[] row : dataBlock.asRowHeap().getRows()) {
-        if (handleRowMatched(row)) {
-          rows.add(row);
-        }
-      }
-      sampleAndCheckInterruption();
-      if (!rows.isEmpty()) {
-        return new RowHeapDataBlock(rows, _dataSchema);
-      }
-    }
-  }
+  protected abstract MseBlock processRightOperator();
 
   @Override
   protected StatMap<?> copyStatMaps() {
     return new StatMap<>(_statMap);
   }
-
-  /**
-   * Returns true if the row is matched.
-   * <p>
-   * Also updates the right row set based on the Operator.
-   *
-   * @param row the row to be checked for matching.
-   * @return true if the row is matched.
-   */
-  protected abstract boolean handleRowMatched(Object[] row);
 
   public enum StatKey implements StatMap.Key {
     //@formatter:off
