+++ pinot-query-runtime/src/test/java/org/apache/pinot/query/runtime/operator/set/IntersectOperatorTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.query.runtime.operator;
+package org.apache.pinot.query.runtime.operator.set;
 
 import com.google.common.collect.ImmutableList;
 import java.util.Arrays;
@@ -26,6 +26,8 @@ import org.apache.pinot.query.routing.VirtualServerAddress;
 import org.apache.pinot.query.runtime.blocks.ErrorMseBlock;
 import org.apache.pinot.query.runtime.blocks.MseBlock;
 import org.apache.pinot.query.runtime.blocks.SuccessMseBlock;
+import org.apache.pinot.query.runtime.operator.MultiStageOperator;
+import org.apache.pinot.query.runtime.operator.OperatorTestUtil;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
@@ -35,7 +37,7 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 
-public class MinusOperatorTest {
+public class IntersectOperatorTest {
   private AutoCloseable _mocks;
 
   @Mock
@@ -60,28 +62,27 @@ public class MinusOperatorTest {
   }
 
   @Test
-  public void testExceptOperator() {
+  public void testIntersectOperator() {
     DataSchema schema = new DataSchema(new String[]{"int_col", "string_col"}, new DataSchema.ColumnDataType[]{
         DataSchema.ColumnDataType.INT, DataSchema.ColumnDataType.STRING
     });
     Mockito.when(_leftOperator.nextBlock())
-        .thenReturn(OperatorTestUtil.block(schema, new Object[]{1, "AA"}, new Object[]{2, "BB"}, new Object[]{3, "CC"},
-            new Object[]{4, "DD"}))
+        .thenReturn(OperatorTestUtil.block(schema, new Object[]{1, "AA"}, new Object[]{2, "BB"}, new Object[]{3, "CC"}))
         .thenReturn(SuccessMseBlock.INSTANCE);
     Mockito.when(_rightOperator.nextBlock()).thenReturn(
-            OperatorTestUtil.block(schema, new Object[]{1, "AA"}, new Object[]{2, "BB"}, new Object[]{5, "EE"}))
+            OperatorTestUtil.block(schema, new Object[]{1, "AA"}, new Object[]{2, "BB"}, new Object[]{4, "DD"}))
         .thenReturn(SuccessMseBlock.INSTANCE);
 
-    MinusOperator minusOperator =
-        new MinusOperator(OperatorTestUtil.getTracingContext(), ImmutableList.of(_leftOperator, _rightOperator),
+    IntersectOperator intersectOperator =
+        new IntersectOperator(OperatorTestUtil.getTracingContext(), ImmutableList.of(_leftOperator, _rightOperator),
             schema);
 
-    MseBlock result = minusOperator.nextBlock();
+    MseBlock result = intersectOperator.nextBlock();
     while (result.isEos()) {
-      result = minusOperator.nextBlock();
+      result = intersectOperator.nextBlock();
     }
     List<Object[]> resultRows = ((MseBlock.Data) result).asRowHeap().getRows();
-    List<Object[]> expectedRows = Arrays.asList(new Object[]{3, "CC"}, new Object[]{4, "DD"});
+    List<Object[]> expectedRows = Arrays.asList(new Object[]{1, "AA"}, new Object[]{2, "BB"});
     Assert.assertEquals(resultRows.size(), expectedRows.size());
     for (int i = 0; i < resultRows.size(); i++) {
       Assert.assertEquals(resultRows.get(i), expectedRows.get(i));
@@ -95,24 +96,23 @@ public class MinusOperatorTest {
     });
     Mockito.when(_leftOperator.nextBlock())
         .thenReturn(OperatorTestUtil.block(schema, new Object[]{1, "AA"}, new Object[]{2, "BB"}, new Object[]{3, "CC"},
-            new Object[]{4, "DD"}, new Object[]{1, "AA"}, new Object[]{2, "BB"}, new Object[]{3, "CC"},
-            new Object[]{4, "DD"}))
+            new Object[]{1, "AA"}, new Object[]{2, "BB"}, new Object[]{3, "CC"}))
         .thenReturn(SuccessMseBlock.INSTANCE);
     Mockito.when(_rightOperator.nextBlock()).thenReturn(
-            OperatorTestUtil.block(schema, new Object[]{1, "AA"}, new Object[]{2, "BB"}, new Object[]{5, "EE"},
-                new Object[]{1, "AA"}, new Object[]{2, "BB"}, new Object[]{5, "EE"}))
+            OperatorTestUtil.block(schema, new Object[]{1, "AA"}, new Object[]{2, "BB"}, new Object[]{4, "DD"},
+                new Object[]{1, "AA"}, new Object[]{2, "BB"}, new Object[]{4, "DD"}))
         .thenReturn(SuccessMseBlock.INSTANCE);
 
-    MinusOperator minusOperator =
-        new MinusOperator(OperatorTestUtil.getTracingContext(), ImmutableList.of(_leftOperator, _rightOperator),
+    IntersectOperator intersectOperator =
+        new IntersectOperator(OperatorTestUtil.getTracingContext(), ImmutableList.of(_leftOperator, _rightOperator),
             schema);
 
-    MseBlock result = minusOperator.nextBlock();
+    MseBlock result = intersectOperator.nextBlock();
     while (result.isEos()) {
-      result = minusOperator.nextBlock();
+      result = intersectOperator.nextBlock();
     }
     List<Object[]> resultRows = ((MseBlock.Data) result).asRowHeap().getRows();
-    List<Object[]> expectedRows = Arrays.asList(new Object[]{3, "CC"}, new Object[]{4, "DD"});
+    List<Object[]> expectedRows = Arrays.asList(new Object[]{1, "AA"}, new Object[]{2, "BB"});
     Assert.assertEquals(resultRows.size(), expectedRows.size());
     for (int i = 0; i < resultRows.size(); i++) {
       Assert.assertEquals(resultRows.get(i), expectedRows.get(i));
@@ -130,13 +130,13 @@ public class MinusOperatorTest {
     Mockito.when(_rightOperator.nextBlock())
         .thenReturn(ErrorMseBlock.fromException(new RuntimeException("Error in right operator")));
 
-    MinusOperator minusOperator =
-        new MinusOperator(OperatorTestUtil.getTracingContext(), ImmutableList.of(_leftOperator, _rightOperator),
+    IntersectOperator intersectOperator =
+        new IntersectOperator(OperatorTestUtil.getTracingContext(), ImmutableList.of(_leftOperator, _rightOperator),
             schema);
-    MseBlock result = minusOperator.nextBlock();
+    MseBlock result = intersectOperator.nextBlock();
     // Keep calling nextBlock until we get an EoS block
     while (!result.isEos()) {
-      result = minusOperator.nextBlock();
+      result = intersectOperator.nextBlock();
     }
     Assert.assertTrue(result.isError());
   }
@@ -152,13 +152,13 @@ public class MinusOperatorTest {
         .thenReturn(OperatorTestUtil.block(schema, new Object[]{3, "aa"}, new Object[]{4, "bb"}))
         .thenReturn(SuccessMseBlock.INSTANCE);
 
-    MinusOperator minusOperator =
-        new MinusOperator(OperatorTestUtil.getTracingContext(), ImmutableList.of(_leftOperator, _rightOperator),
+    IntersectOperator intersectOperator =
+        new IntersectOperator(OperatorTestUtil.getTracingContext(), ImmutableList.of(_leftOperator, _rightOperator),
             schema);
-    MseBlock result = minusOperator.nextBlock();
+    MseBlock result = intersectOperator.nextBlock();
     // Keep calling nextBlock until we get an EoS block
     while (!result.isEos()) {
-      result = minusOperator.nextBlock();
+      result = intersectOperator.nextBlock();
     }
     Assert.assertTrue(result.isError());
   }
