+++ pinot-query-runtime/src/test/java/org/apache/pinot/query/runtime/operator/set/MinusAllOperatorTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.query.runtime.operator;
+package org.apache.pinot.query.runtime.operator.set;
 
 import com.google.common.collect.ImmutableList;
 import java.util.Arrays;
@@ -25,6 +25,8 @@ import org.apache.pinot.common.utils.DataSchema;
 import org.apache.pinot.query.routing.VirtualServerAddress;
 import org.apache.pinot.query.runtime.blocks.MseBlock;
 import org.apache.pinot.query.runtime.blocks.SuccessMseBlock;
+import org.apache.pinot.query.runtime.operator.MultiStageOperator;
+import org.apache.pinot.query.runtime.operator.OperatorTestUtil;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
@@ -34,8 +36,7 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 
-public class IntersectAllOperatorTest {
-
+public class MinusAllOperatorTest {
   private AutoCloseable _mocks;
 
   @Mock
@@ -60,27 +61,27 @@ public class IntersectAllOperatorTest {
   }
 
   @Test
-  public void testIntersectAllOperator() {
-    DataSchema schema = new DataSchema(new String[]{"int_col"},
-        new DataSchema.ColumnDataType[]{DataSchema.ColumnDataType.INT});
-
+  public void testExceptAllOperator() {
+    DataSchema schema = new DataSchema(new String[]{"int_col"}, new DataSchema.ColumnDataType[]{
+        DataSchema.ColumnDataType.INT, DataSchema.ColumnDataType.STRING
+    });
     Mockito.when(_leftOperator.nextBlock())
-        .thenReturn(OperatorTestUtil.block(schema, new Object[]{1}, new Object[]{2}, new Object[]{3}))
+        .thenReturn(OperatorTestUtil.block(schema, new Object[]{1}, new Object[]{2}, new Object[]{3}, new Object[]{4}))
         .thenReturn(SuccessMseBlock.INSTANCE);
     Mockito.when(_rightOperator.nextBlock()).thenReturn(
-            OperatorTestUtil.block(schema, new Object[]{1}, new Object[]{2}, new Object[]{4}))
+            OperatorTestUtil.block(schema, new Object[]{1}, new Object[]{2}, new Object[]{5}))
         .thenReturn(SuccessMseBlock.INSTANCE);
 
-    IntersectAllOperator intersectOperator =
-        new IntersectAllOperator(OperatorTestUtil.getTracingContext(), ImmutableList.of(_leftOperator, _rightOperator),
+    MinusAllOperator minusOperator =
+        new MinusAllOperator(OperatorTestUtil.getTracingContext(), ImmutableList.of(_leftOperator, _rightOperator),
             schema);
 
-    MseBlock result = intersectOperator.nextBlock();
+    MseBlock result = minusOperator.nextBlock();
     while (result.isEos()) {
-      result = intersectOperator.nextBlock();
+      result = minusOperator.nextBlock();
     }
     List<Object[]> resultRows = ((MseBlock.Data) result).asRowHeap().getRows();
-    List<Object[]> expectedRows = Arrays.asList(new Object[]{1}, new Object[]{2});
+    List<Object[]> expectedRows = Arrays.asList(new Object[]{3}, new Object[]{4});
     Assert.assertEquals(resultRows.size(), expectedRows.size());
     for (int i = 0; i < resultRows.size(); i++) {
       Assert.assertEquals(resultRows.get(i), expectedRows.get(i));
@@ -88,28 +89,28 @@ public class IntersectAllOperatorTest {
   }
 
   @Test
-  public void testIntersectAllOperatorWithDups() {
-    DataSchema schema = new DataSchema(new String[]{"int_col"},
-        new DataSchema.ColumnDataType[]{DataSchema.ColumnDataType.INT});
-
+  public void testExceptAllOperatorWithDups() {
+    DataSchema schema = new DataSchema(new String[]{"int_col"}, new DataSchema.ColumnDataType[]{
+        DataSchema.ColumnDataType.INT, DataSchema.ColumnDataType.STRING
+    });
     Mockito.when(_leftOperator.nextBlock())
-        .thenReturn(OperatorTestUtil.block(schema, new Object[]{1}, new Object[]{2}, new Object[]{2}, new Object[]{3},
-            new Object[]{3}, new Object[]{3}))
+        .thenReturn(OperatorTestUtil.block(schema, new Object[]{1}, new Object[]{2}, new Object[]{2}, new Object[]{2},
+            new Object[]{3}, new Object[]{3}, new Object[]{3}))
         .thenReturn(SuccessMseBlock.INSTANCE);
     Mockito.when(_rightOperator.nextBlock()).thenReturn(
             OperatorTestUtil.block(schema, new Object[]{2}, new Object[]{3}, new Object[]{3}, new Object[]{4}))
         .thenReturn(SuccessMseBlock.INSTANCE);
 
-    IntersectAllOperator intersectOperator =
-        new IntersectAllOperator(OperatorTestUtil.getTracingContext(), ImmutableList.of(_leftOperator, _rightOperator),
+    MinusAllOperator minusOperator =
+        new MinusAllOperator(OperatorTestUtil.getTracingContext(), ImmutableList.of(_leftOperator, _rightOperator),
             schema);
 
-    MseBlock result = intersectOperator.nextBlock();
+    MseBlock result = minusOperator.nextBlock();
     while (result.isEos()) {
-      result = intersectOperator.nextBlock();
+      result = minusOperator.nextBlock();
     }
     List<Object[]> resultRows = ((MseBlock.Data) result).asRowHeap().getRows();
-    List<Object[]> expectedRows = Arrays.asList(new Object[]{2}, new Object[]{3}, new Object[]{3});
+    List<Object[]> expectedRows = Arrays.asList(new Object[]{1}, new Object[]{2}, new Object[]{2}, new Object[]{3});
     Assert.assertEquals(resultRows.size(), expectedRows.size());
     for (int i = 0; i < resultRows.size(); i++) {
       Assert.assertEquals(resultRows.get(i), expectedRows.get(i));
